requirements.txt
pandas>=2.2.2
oracledb>=2.2.1
python-dotenv>=0.19.1
